Refactor project structure and update .gitignore
This commit introduces significant restructuring of the project, including the removal of ArithmeticCalculator.java, ArithmeticOperationsTest.java and workspace.xml files. Additionally, various .iml files are added, and the .gitignore file is updated to better match the project's needs. A few source codes are also lightly modified as part of the clean-up process.