Implement draw method overload and shape polymorphism
Overloaded the draw method in the Artist class, allowing it to accept a Shape object or a String as a parameter. Added the InactiveUser class extending User with overridden setUsername method. Introduced BrokenCircle and BrokenRedCircle classes to exemplify polymorphism and method overloading. Changing Shape from abstract to concrete class along with corresponding changes in Circle class.